Add confirmation popup

+++ Vue_3_Traversy_Media/2nd/vue-jobs-new/src/views/JobView.vue
@@ -29,6 +29,8 @@ onMounted(async () => {
 
 const deleteJob = async () => {
   try {
+    const confirm = window.confirm("Are you sure you want to delete this job?");
+    if (!confirm) return;
     await axios.delete(`/api/jobs/${jobId}`);
     router.push("/jobs");
     notifications.success('Job deleted successfully');
